requirement.txt
numpy
pandas
plotly
matplotlib
seaborn
flask
Flask-SQLAlchemy
tkcalendar
scikit-learn
pypokedex
pillow
flet
nltk
textblob
newspaper3k
pygame
email-to
ttkbootstrap
Faker
yfinance
streamlit
colorthief
fastapi
uvicorn
python-multipart
opencv-python
tensorflow
fpdf
validators
tkintermapview
pyspark
findspark
win10toast
pyjokes
langdetect
langcodes
forex_python
python-vlc
rembg
pynput
keyboard
windows-curses
docx2pdf
networkx
transformers
speedtest-cli
stegano
Werkzeug
torch
tf-keras
folium
pynacl
discord 
yt_dlp
pycountry
kivy
openpyxl
xlrd
pyodbc